profile local storage

+++ src/Components/Pages/Car.js
@@ -3,33 +3,41 @@ import { useFirebaseApp, useUser } from "reactfire";
 import { db } from "../../firebaseProyect";
 
 const Car = () => {
-    const { data: user } = useUser();
+  const { data: user } = useUser();
   const [cards, setCards] = useState([]);
 
-  const fetchData = async () => {    
-    let uID = user.uid;
-      const data = await db.collection(uID).get();
+  const fetchData = async () => {
+    if (user) {
+      let uID = user.uid;
+      const data = await db.collection("car-"+uID).get();
       setCards(data.docs.map((doc) => ({ ...doc.data(), id: doc.id })));
-    };
+    }
+  };
 
-  useEffect(() => {    
+  useEffect(() => {
     fetchData();
   }, []);
 
   const onDelete = async (product) => {
-    let uID = user.uid;
-    await db.collection(uID).doc(product.id).delete();
-    fetchData();
+    if (user) {
+      let uID = user.uid;
+      await db.collection(uID).doc(product.id).delete();
+      fetchData();
+    }
   };
 
   return (
-    <div className="card-view text-center">
+    <div className="card-view text-center mt-5">
       <div className="row">
         {cards.map((product) => {
           return (
             <div className="col">
               <div className="card">
-                <img src={product.thumbnail} className="card-img-top" alt="..." />
+                <img
+                  src={product.thumbnail}
+                  className="card-img-top"
+                  alt="..."
+                />
                 <div className="card-body">
                   <h5 className="card-title">{product.name}</h5>
                   <p className="card-text">${product.price}</p>
